requirements.txt
captum==0.5.0
foolbox==3.3.3
matplotlib==3.5.2
numpy==1.22.3
PyYAML==6.0
torch==1.11.0
torchvision==0.12.0
